+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   limited to.
 
 * Breaking change:
+
   Following stat creation methods are removed:
     - `StatefulService#getTimeSeriesStat`
     - `StatefulService#getTimeSeriesHistogramStat`
@@ -29,6 +30,18 @@
   "[RemoteHostID]GossipPatchDurationMicrosPerDay". However, it was actually an hourly stat.
   So, it was renamed to "[RemoteHostID]GossipPatchDurationMicrosPerHour"
 
+* Breaking change: Gson's '@SerializedName(value)' is now properly supported,
+  the field name exposed in the DocumentDescription and available in xenon
+  queries now respects Gson's @SerializedName annotation.
+
+  Previously, the exposed JSON document respected this annotation but the
+  DocumentDescription and query service did not respect the annotation,
+  exposing the underlying field with its java name, which was incoherent.
+
+  As a result of this fields which had the 'SerializeName' annotation
+  may no longer be properly indexed as the field name has changed.
+
+  See: https://www.pivotaltracker.com/story/show/148644281  
 
 ## 1.5.1